Add a metric ton of compiler warnings to the Xcode project, and clean up the source accordingly.
At some point these warnings should be merged into the makefile system,
but I don't have the will to merge them into the 3451351235 individual
makefiles at the moment.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@12 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/items.cc
@@ -1837,7 +1837,7 @@ static std::string drop_selitem_text( const std::vector<MenuEntry*> *s )
         }
     }
     
-    snprintf( buf, sizeof buf, "  (%u%s turn%s)", s->size(),
+    snprintf( buf, sizeof buf, "  (%lu%s turn%s)", (unsigned long) (s->size()),
               extraturns? "+" : "",
               s->size() > 1? "s" : "" );
     return buf;
